tutor/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
boto3>=1.34.0
botocore>=1.34.0
pyyaml>=6.0
strands-agents>=1.37.0
aws-bedrock-token-generator>=0.1.0
openai>=1.0.0
numpy>=1.24.0
pillow>=10.0.0
